% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.project-package.R
\docType{package}
\name{CI.project-package}
\alias{CI.project}
\alias{CI.project-package}
\title{CI.project: Housing Data For Luxembourg}
\description{
This package contains functions to get, clean and analyse housing price data for Luxembourg.
}
\author{
\strong{Maintainer}: Bruno Rodrigues \email{bruno@brodrigues.co} (\href{https://orcid.org/0000-0002-3211-3689}{ORCID})

}
\keyword{internal}
