% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{clean_raw_data}
\alias{clean_raw_data}
\title{clean_raw_data Cleans the raw data}
\usage{
clean_raw_data(raw_data)
}
\arguments{
\item{raw_data}{The raw data to clean (the output of get_raw_data)}
}
\value{
A data frame
}
\description{
clean_raw_data Cleans the raw data
}
\details{
Removes uneeded rows, renames localities to make their names consistent across years and converts columns with prices to numeric columns
}
