% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{commune_level_data}
\alias{commune_level_data}
\title{commune_level_data}
\format{
A data frame with 1171 rows and 5 variables:
\describe{
\item{ year }{ 2010 to 2020 }
\item{ locality }{ Communes of Luxembourg }
\item{ n_offers }{ Total offers in that commune for that year }
\item{ average_price_nominal_euros }{ Average selling price of houses in that commune for that year }
\item{ average_price_m2_nominal_euros }{ Average selling price by m2 of houses in that commune for that year }
}
}
\source{
Observatoire de l'Habitat
}
\usage{
commune_level_data
}
\description{
data in 'luHousePrices'.
}
\keyword{datasets}
