% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{communes}
\alias{communes}
\title{communes}
\format{
A list of luxembourguish communes
}
\source{
Observatoire de l'Habitat
}
\usage{
communes
}
\description{
data in 'luHousePrices'.
}
\keyword{datasets}
