% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{country_level_data}
\alias{country_level_data}
\title{country_level_data}
\format{
A data frame with 11 rows and 5 variables:
\describe{
\item{ year }{ 2010 to 2020 }
\item{ locality }{ For this dataset, only the country of Luxembourg }
\item{ n_offers }{ Total offers in the country for that year }
\item{ average_price_nominal_euros }{ Average selling price of houses in the commune for that year }
\item{ average_price_m2_nominal_euros }{ Average selling price by m2 of houses in the commune for that year }
}
}
\source{
Observatoire de l'Habitat
}
\usage{
country_level_data
}
\description{
data in 'luHousePrices'.
}
\keyword{datasets}
