% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_former_communes}
\alias{get_former_communes}
\title{get_former_communes Downloads list of former communes from Wikipedia}
\usage{
get_former_communes(
  url = "https://en.wikipedia.org/wiki/Communes_of_Luxembourg#Former_communes",
  min_year = 2009,
  table_position = 3
)
}
\arguments{
\item{url}{Optional: Persistent url to the data}

\item{min_year}{Optional: Minimum year to consider. Defaults to 2009 because price data starts in 2010}

\item{table_position}{Optional: Scraping returns a list of tables, so users need to specify the correct table. Defaults to 3, the position of the table as of writing.}
}
\value{
A data frame
}
\description{
get_former_communes Downloads list of former communes from Wikipedia
}
