% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_test_communes}
\alias{get_test_communes}
\title{get_test_communes Creates list of communes that should be in the data}
\usage{
get_test_communes(former_communes, current_communes)
}
\arguments{
\item{former_communes}{Former communes df as returned by get_former_communes()}

\item{current_communes}{Current communes df as returned by get_current_communes()}
}
\value{
A data frame
}
\description{
get_test_communes Creates list of communes that should be in the data
}
